python/pyproject.toml
[project]
name = "prompt-weave-core"
version = "0.1.0"
description = "Core logic for the prompt-weave VS Code extension"
requires-python = ">=3.9"
dependencies = [
    "python-frontmatter>=1.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["prompt_weave"]

[tool.pytest.ini_options]
testpaths = ["tests"]
